fix conflicting types

+++ math/tests/sqrt.test.cpp
@@ -75,7 +75,8 @@ TEST(sqrt_function, sqrt_float)
     static constexpr float f{4.2375f};
     static constexpr float fConsteig{consteig::sqrt(f)};
     static_assert(
-        consteig::compareFloats(fConsteig, 2.0585189f, CONSTEIG_TEST_TOLERANCE),
+        consteig::compareFloats(fConsteig, 2.0585189f,
+                                static_cast<float>(CONSTEIG_TEST_TOLERANCE)),
         MSG);
     ASSERT_NEAR(fConsteig, 2.0585189f, CONSTEIG_TEST_TOLERANCE);
 }
@@ -84,9 +85,10 @@ TEST(sqrt_function, csqrt_positive)
 {
     static constexpr float g{22.2f};
     static constexpr Complex<float> gConsteig{consteig::csqrt(g)};
-    static_assert(consteig::compareFloats(gConsteig.real, 4.711687f,
-                                          CONSTEIG_TEST_TOLERANCE),
-                  MSG);
+    static_assert(
+        consteig::compareFloats(gConsteig.real, 4.711687f,
+                                static_cast<float>(CONSTEIG_TEST_TOLERANCE)),
+        MSG);
     static_assert(gConsteig.imag == 0.0f, MSG);
     ASSERT_NEAR(gConsteig.real, 4.711687f, CONSTEIG_TEST_TOLERANCE);
     ASSERT_EQ(gConsteig.imag, 0.0f);
@@ -97,9 +99,10 @@ TEST(sqrt_function, csqrt_negative)
     static constexpr float g{-22.2f};
     static constexpr Complex<float> gConsteig{consteig::csqrt(g)};
     static_assert(gConsteig.real == 0.0f, MSG);
-    static_assert(consteig::compareFloats(gConsteig.imag, 4.711687f,
-                                          CONSTEIG_TEST_TOLERANCE),
-                  MSG);
+    static_assert(
+        consteig::compareFloats(gConsteig.imag, 4.711687f,
+                                static_cast<float>(CONSTEIG_TEST_TOLERANCE)),
+        MSG);
     ASSERT_EQ(gConsteig.real, 0.0f);
     ASSERT_NEAR(gConsteig.imag, 4.711687f, CONSTEIG_TEST_TOLERANCE);
